repaint/requirements.txt

# blas==1.0
blobfile==2.0.2
blosc==1.21.3
boost-cpp==1.82.0
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2023.08.22
cairo==1.16.0
certifi==2023.11.17
cffi==1.15.1
cfitsio==3.470
charset-normalizer==2.0.4
contourpy==1.0.5
cryptography==41.0.3
cuda-cudart==11.7.99
cuda-cupti==11.7.101
cuda-libraries==11.7.1
cuda-nvrtc==11.7.99
cuda-nvtx==11.7.91
cuda-runtime==11.7.1
cycler==0.11.0
cyrus-sasl==2.1.28
dbus==1.13.18
expat==2.5.0
ffmpeg==4.3
filelock==3.12.2
fontconfig==2.14.1
fonttools==4.25.0
freetype==2.12.1
freexl==1.0.6
ftfy==6.1.1
gdal==3.6.2
geos==3.8.0
geotiff==1.7.0
giflib==5.2.1
glib==2.69.1
gmp==6.2.1
gnutls==3.6.15
gst-plugins-base==1.14.1
gstreamer==1.14.1
hdf4==4.2.13
hdf5==1.12.1
icu==73.1
idna==3.4
imageio==2.34.0
importlib_resources==6.1.0
intel-openmp==2023.1.0
# jpeg==9e
json-c==0.16
kealib==1.5.0
kiwisolver==1.4.4
krb5==1.20.1
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libboost==1.82.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libclang==14.0.6
libclang13==14.0.6
libcublas==11.10.3.66
libcufft==10.7.2.124
libcufile==1.7.0.149
libcups==2.4.2
libcurand==10.3.3.53
libcurl==8.2.1
libcusolver==11.4.0.1
libcusparse==11.7.4.91
libdeflate==1.17
libedit==3.1.20221030
libev==4.33
libffi==3.4.4
libgcc-ng==11.2.0
libgdal==3.6.2
libgfortran-ng==7.5.0
libgfortran4==7.5.0
libgomp==11.2.0
libiconv==1.16
libidn2==2.3.4
libkml==1.3.0
libllvm14==14.0.6
libnetcdf==4.8.1
libnghttp2==1.52.0
libnpp==11.7.4.75
libnvjpeg==11.8.0.2
libpng==1.6.39
libpq==12.15
libspatialite==4.3.0a
libssh2==1.10.0
libstdcxx-ng==11.2.0
libtasn1==4.19.0
libtiff==4.5.0
libunistring==0.9.10
libuuid==1.41.5
libwebp==1.2.4
libwebp-base==1.2.4
libxcb==1.15
libxkbcommon==1.0.1
libxml2==2.10.4
libzip==1.8.0
lmdb==1.4.1
lxml==4.9.2
lz4-c==1.9.4
matplotlib==3.7.2
matplotlib-base==3.7.2
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.6
mkl_random==1.2.2
mpi==1.0
mpi4py==3.1.4
mpich==3.3.2
munkres==1.1.4
mysql==5.7.24
ncurses==6.4
nettle==3.7.3
nspr==4.35
nss==3.89.1
numpy==1.24.3
numpy-base==1.24.3
opencv-python==4.8.0.74
openh264==2.1.1
openjpeg==2.4.0
# openssl==1.1.1w
packaging==23.1
pcre==8.45
pcre2==10.42
pillow==9.4.0
pip==23.1.2
pixman==0.40.0
ply==3.11
poppler==22.12.0
poppler-data==0.4.11
proj==6.2.1
pycparser==2.21
pycryptodomex==3.18.0
pyopenssl==23.2.0
pyparsing==3.0.9
pyqt==5.15.10
pyqt5-sip==12.13.0
pysocks==1.7.1
python==3.8.18
python-dateutil==2.8.2
pytorch==1.13.0
pytorch-cuda==11.7
pytorch-mutex==1.0
pyyaml==6.0
qhull==2020.2
qt-main==5.15.2
readline==8.2
requests==2.29.0
setuptools==67.8.0
sip==6.7.12
six==1.16.0
sqlite==3.41.2
tbb==2021.8.0
tifffile==2023.4.12
tiffile==2018.10.18
tiledb==2.3.3
tk==8.6.12
tomli==2.0.1
torchaudio==0.13.0
torchsummary==1.5.1
torchvision==0.14.0
tornado==6.3.3
tqdm==4.65.0
typing_extensions==4.6.3
uriparser==0.9.7
urllib3==1.26.16
wcwidth==0.2.6
wheel==0.38.4
xerces-c==3.2.4
xz==5.4.2
yaml==0.2.5
zipp==3.11.0
zlib==1.2.13
zstd==1.5.5
